+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBDatabase.java
@@ -336,7 +336,7 @@ public class FDBDatabase {
                 // otherwise getReadVersion does not use the cached value and results in a GRV call to FDB
                 if (version >= weakReadSemantics.getMinVersion() &&
                         (System.currentTimeMillis() - versionTimeMillis) <= weakReadSemantics.getStalenessBoundMillis()) {
-                    context.ensureActive().setReadVersion(version);
+                    context.setReadVersion(version);
                     if (timer != null) {
                         timer.increment(FDBStoreTimer.Counts.SET_READ_VERSION_TO_LAST_SEEN);
                     }
@@ -407,6 +407,9 @@ public class FDBDatabase {
         try {
             // Set the read version of the transaction, then read it back. This requires no I/O, but it does
             // require the network thread be running. The exact value used for the read version is unimportant.
+            // Note that this calls setReadVersion and getReadVersion on the Transaction object, *not* on the
+            // FDBRecordContext. This is because the FDBRecordContext will cache the value of setReadVersion to
+            // avoid having to go back to the FDB network thread, but we do not want that for instrumentation.
             final Transaction tr = context.ensureActive();
             final long startTime = System.nanoTime();
             tr.setReadVersion(1066L);
@@ -496,18 +499,12 @@ public class FDBDatabase {
      * Measuring it explicitly gives an indication of the cluster's GRV latency, which is driven by its rate keeping.
      * @param context transaction to use to access the database
      * @return a future that will be completed with the read version of the current transaction
+     * @deprecated use {@link FDBRecordContext#getReadVersionAsync()} instead
      */
+    @Deprecated
+    @Nonnull
     public CompletableFuture<Long> getReadVersion(@Nonnull FDBRecordContext context) {
-        CompletableFuture<Long> readVersionFuture = injectLatency(FDBLatencySource.GET_READ_VERSION).thenCompose(ignore ->
-                context.ensureActive().getReadVersion());
-        if (!isTrackLastSeenVersionOnRead()) {
-            return readVersionFuture;
-        }
-        long startTime = System.currentTimeMillis();
-        return readVersionFuture.thenApply((Long readVersion) -> {
-            updateLastSeenFDBVersion(startTime, readVersion);
-            return readVersion;
-        });
+        return context.getReadVersionAsync();
     }
 
     // Update lastSeenFDBVersion if readVersion is newer