+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/SizeStatisticsCollectorTest.java
@@ -224,7 +224,7 @@ public class SizeStatisticsCollectorTest extends FDBRecordStoreTestBase {
         try (FDBRecordContext context = openContext()) {
             // Somewhat leaky, but should get transaction_too_old.
             // One can set knob_max_read_transaction_life_versions up to max_versions_in_flight, which is 100000000.
-            context.ensureActive().setReadVersion(commitVersion - 101_000_000);
+            context.setReadVersion(commitVersion - 101_000_000);
             assertThat(statisticsCollector.collect(context, ExecuteProperties.SERIAL_EXECUTE), is(false));
             assertThrows(FDBExceptions.FDBStoreTransactionIsTooOldException.class, context::commit);
         }